requirements.txt
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
flasgger==0.9.7.1
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.1
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
mistune==3.1.3
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
XlsxWriter==3.2.2
